+++ src/hooks/useFetchData.js
@@ -4,31 +4,28 @@ import constants from '../data/constants'
 
 const useFetchData = (initialValue = []) => {
   const { api } = constants
-  const [postData, setPostData] = useState(initialValue)
+  const [fetchData, setFetchData] = useState(initialValue)
   const [error, setError] = useState(null)
 
   useEffect(() => {
     fetch(api)
       .then(response => response.json())
       .then(json => {
-        setPostData(json)
-        // setError(() => {
-        //   throw new Error('Error!')
-        // })
+        setFetchData(json)
         if (error) {
           setError(null)
         }
       })
       .catch(err => {
         setError(err)
-        // setError(() => {
-        //   throw new Error('Error!')
-        // })
+        setError(() => {
+          throw new Error(err)
+        })
       })
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [])
 
-  return { postData, error }
+  return { fetchData, error }
 }
 
 export default useFetchData
